// src/exchanges/bitfinex/enums/BitfinexOrderTypesEnum.ts
export enum BitfinexOrderTypesEnum {
  MARKET = 'MARKET',
  LIMIT = 'LIMIT',
  STOP = 'STOP',
  STOP_LIMIT = 'STOP_LIMIT',
  TRAILING_STOP = 'TRAILING_STOP',
  FOK = 'FOK',
  IOC = 'IOC',
  EXCHANGE_MARKET = 'EXCHANGE_MARKET',
  EXCHANGE_LIMIT = 'EXCHANGE_LIMIT',
  EXCHANGE_STOP = 'EXCHANGE_STOP',
  EXCHANGE_STOP_LIMIT = 'EXCHANGE_STOP_LIMIT',
  EXCHANGE_TRAILING_STOP = 'EXCHANGE_TRAILING_STOP',
  EXCHANGE_FOK = 'EXCHANGE_FOK',
  EXCHANGE_IOC = 'EXCHANGE_IOC',
}